refactor(cobol): extract helpers
