+++ compile/cobol/compiler.go
@@ -40,68 +40,6 @@ func New(env *types.Env) *Compiler {
 	}
 }
 
-func (c *Compiler) writeIndent() {
-	for i := 0; i < c.indent; i++ {
-		c.buf.WriteString("    ")
-	}
-}
-
-func (c *Compiler) writeln(s string) {
-	c.writeIndent()
-	c.buf.WriteString(s)
-	c.buf.WriteByte('\n')
-}
-
-func (c *Compiler) newTemp() string {
-	name := fmt.Sprintf("TMP%d", c.tmpCounter)
-	c.tmpCounter++
-	return name
-}
-
-func cobolName(name string) string {
-	up := strings.ToUpper(name)
-	switch up {
-	case "END", "START":
-		return "V_" + up
-	}
-	return up
-}
-
-func isSimpleExpr(n *ast.Node) bool {
-	switch n.Kind {
-	case "int", "float", "selector", "string", "bool":
-		return true
-	case "unary":
-		if n.Value == "-" {
-			return isSimpleExpr(n.Children[0])
-		}
-	case "group":
-		return isSimpleExpr(n.Children[0])
-	}
-	return false
-}
-
-// picForType returns a picture clause for the given static type.
-func (c *Compiler) picForType(t types.Type) string {
-	switch t.(type) {
-	case types.StringType:
-		return "PIC X(100)."
-	case types.FloatType:
-		return "PIC 9(4)V9(4)."
-	}
-	return "PIC 9."
-}
-
-// declare records a WORKING-STORAGE declaration.
-func (c *Compiler) declare(line string) {
-	for _, d := range c.decls {
-		if d == line {
-			return
-		}
-	}
-	c.decls = append(c.decls, line)
-}
-
 // Compile translates prog into COBOL source code.
 // Only a very small subset of the language is supported.
 func (c *Compiler) Compile(prog *parser.Program) ([]byte, error) {
@@ -1127,110 +1065,6 @@ func (c *Compiler) compileMatchExpr(n *ast.Node) string {
 		c.writeln("    END-IF")
 	}
 	return resTmp
-}
-
-func extractInt(n *ast.Node) int {
-	if n.Kind == "int" {
-		switch v := n.Value.(type) {
-		case int:
-			return v
-		case float64:
-			return int(v)
-		}
-	}
-	return 0
-}
-
-// listInfo extracts information about a list expression. It returns the COBOL
-// variable name if the list is stored in a variable, the list length, the
-// element picture clause and the literal elements if any.
-func (c *Compiler) listInfo(n *ast.Node) (name string, length int, pic string, elems []*ast.Node, ok bool) {
-	switch n.Kind {
-	case "list":
-		length = len(n.Children)
-		if length > 0 {
-			pic = c.picForExpr(n.Children[0])
-		} else {
-			pic = "PIC 9."
-		}
-		elems = n.Children
-		ok = true
-		return
-	case "selector":
-		if l, ok2 := c.listLens[n.Value.(string)]; ok2 {
-			length = l
-			ok = true
-		}
-		if c.env != nil {
-			if t, err := c.env.GetVar(n.Value.(string)); err == nil {
-				if lt, ok2 := t.(types.ListType); ok2 {
-					if pic == "" {
-						pic = c.picForType(lt.Elem)
-					}
-					ok = true
-				}
-			}
-		}
-		name = cobolName(n.Value.(string))
-		if pic == "" {
-			pic = "PIC 9."
-		}
-		return
-	}
-	return
-}
-
-// concatLists emits COBOL code to concatenate two lists and returns the result
-// variable name.
-func (c *Compiler) concatLists(left, right *ast.Node) string {
-	ln, llen, lpic, lelems, ok1 := c.listInfo(left)
-	rn, rlen, rpic, relems, ok2 := c.listInfo(right)
-	if !ok1 || !ok2 {
-		c.writeln("    *> unsupported list concatenation")
-		return "0"
-	}
-	pic := lpic
-	if pic == "" {
-		pic = rpic
-	}
-	res := c.newTemp()
-	c.declare(fmt.Sprintf("01 %s OCCURS %d TIMES %s", res, llen+rlen, pic))
-	c.listLens[res] = llen + rlen
-	idx := 1
-	if ln != "" {
-		for i := 0; i < llen; i++ {
-			c.writeln(fmt.Sprintf("    MOVE %s(%d) TO %s(%d)", ln, i+1, res, idx))
-			idx++
-		}
-	} else {
-		for _, ch := range lelems {
-			c.writeln(fmt.Sprintf("    MOVE %s TO %s(%d)", c.expr(ch), res, idx))
-			idx++
-		}
-	}
-	if rn != "" {
-		for i := 0; i < rlen; i++ {
-			c.writeln(fmt.Sprintf("    MOVE %s(%d) TO %s(%d)", rn, i+1, res, idx))
-			idx++
-		}
-	} else {
-		for _, ch := range relems {
-			c.writeln(fmt.Sprintf("    MOVE %s TO %s(%d)", c.expr(ch), res, idx))
-			idx++
-		}
-	}
-	return res
-}
-
-func extractIntList(n *ast.Node) []int {
-	if n.Kind != "list" {
-		return nil
-	}
-	res := make([]int, 0, len(n.Children))
-	for _, ch := range n.Children {
-		res = append(res, extractInt(ch))
-	}
-	return res
 }
 
 func (c *Compiler) compileLet(n *ast.Node) {
